==> 02_content_recommender.ipynb <==
# imports
import pandas as pd
from openai.embeddings_utils import distances_from_embeddings

#upload data with user interactions
users = pd.read_csv("embeddings/users_emb_TEST.csv") #document with user interactions
users.columns =['User', 'ID', 'Interactions_emb']
users.head()

#upload the data with news articles
news = pd.read_csv("embeddings/news_emb_TEST.csv") #document with user interactions
news = news.drop(['Content', 'Category', 'SubCategory'], axis=1)
news.head()

# Create a dictionary with user interactions
user_dict = {}
for index, row in users.iterrows():
    user = row['User']
    interactions = row['Interactions_emb']
    user_dict[user] = ('Content', interactions)

for user, (content, embeddings_str) in user_dict.items():
    # Check for empty string before converting to float
    embeddings_list = [float(value) if value.strip() else 0.0 for value in embeddings_str.strip('[]').split(',')]
    user_dict[user] = (content, embeddings_list)


# Create a dictionary with news interactions
news_dict = {}
for index, row in news.iterrows():
    news_id = row['ID']
    interactions = row['Content_emb']
    news_dict[news_id] = ('Content', interactions)

for news_id, (content, embeddings_str) in news_dict.items():
    # Check for empty string before converting to float
    embeddings_list = [float(value) if value.strip() else 0.0 for value in embeddings_str.strip('[]').split(',')]
    news_dict[news_id] = (content, embeddings_list)

#create list with interacted items for a particular user
def user_list(user_id, df):
    user_list = df.loc[df['User'] == user_id, 'ID'].tolist()
    if len(user_list) > 0:
        user_id_str = str(user_list[0])  # Convert to string
        return user_id_str.split()
    else:
        return []

#create content recommendations
def distances_based_on_content(user_id: str):
    user_embedding = user_dict[user_id][1]  # Get the embedding of the user
    news_ids = list(news_dict.keys()) #list of news articles
    news_embeddings = [news_dict[news_id][1] for news_id in news_ids]# Get the embedding of each news article

    # Calculate distances from the user to each news article
    distances = distances_from_embeddings(user_embedding, news_embeddings, distance_metric="cosine")

    # Create a DataFrame with news IDs and corresponding distances
    df = pd.DataFrame({'ID': news_ids, 'Distance': distances})
    sorted_df = df.sort_values(by='Distance', ascending=True)
    
    #delete articles already read by a user
    items_to_delete = user_list(user_id, users)
    filtered_df = sorted_df[~sorted_df['ID'].isin(items_to_delete)]
    
    #reset indexes in a new df 
    filtered_df.reset_index(drop=True, inplace=True)
    
    return filtered_df

distances_based_on_content('U13740')

Saving recommendations for each user

# Assuming you have a DataFrame named 'users' with a 'User' column
users_list = users['User'].tolist()

def list_ids_in_folder(folder_path):
    ids = set()
    
    for filename in os.listdir(folder_path):
        if filename.endswith(".csv"):
            # Extracting the ID from the filename
            file_id = filename.split("_")[0][1:]
            ids.add(file_id)
    
    return list(ids)


for i in users_list:
    #Example usage
    recommendations = distances_based_on_content(user_id=i)
    recommendations.to_csv('content_recommendations/' + i + '_content.csv', index=False)
