==> 03_hybrid_recommender.ipynb <==
# imports
import pandas as pd
import os

def print_hybrid_recommendations (user_id):
    collab_rec = pd.read_csv("collaborative_recommendations_test/" + user_id + "_collab.csv")
    content_rec = pd.read_csv("content_recommendations_test/" + user_id + "_content.csv")
    # Create a dictionary mapping IDs to new column values from df2
    
    merged_df = pd.merge(collab_rec, content_rec, on='ID', how='inner')
    merged_df['hybrid'] = merged_df[['article_distance', 'Distance']].mean(axis=1)

    hybrid_df = merged_df.drop(['article_distance', 'Distance'], axis=1)
    
    return hybrid_df

print_hybrid_recommendations("U13740")

Saving recommendations for each user

def list_ids_in_folder(folder_path):
    ids = set()
    
    for filename in os.listdir(folder_path):
        if filename.endswith(".csv"):
            # Extracting the ID from the filename
            file_id = filename.split("_")[0][1:]
            ids.add(file_id)
    
    ids = ['U' + num for num in ids]       
    
    return list(ids)

folder_path = 'content_recommendations_test'
users_list = list_ids_in_folder(folder_path)

for i in users_list:
    try:
        recommendations = print_hybrid_recommendations(user_id=i)
        recommendations.to_csv('hybrid_recommendations/' + i + '_hybrid.csv', index=False)
    except pd.errors.EmptyDataError:
        print(f"No data found for user {i}. Skipping...")
